fix ssr bugs
